+++ database/migrations/2020_12_22_180025_create_sertifikasi_tuks_table.php
@@ -17,9 +17,9 @@ class CreateSertifikasiTuksTable extends Migration
             $table->bigIncrements('id');
             $table->bigInteger('sertifikasi_id');
             $table->bigInteger('tuk_id');
-            $table->decimal('tuk_price_baru', 12);
-            $table->decimal('tuk_price_perpanjang', 12);
-            $table->decimal('tuk_price_training')->nullable();
+            $table->float('tuk_price_baru', 12);
+            $table->float('tuk_price_perpanjang', 12);
+            $table->float('tuk_price_training')->nullable();
             $table->timestamps();
         });
     }
